MIC-2171: Added worker log timing messages and prototype cli
This change adds additional timing messages in the worker logs for
setup, simulant initialization, etc., and a final JSON-formatted
telemetry string to be logged for each simulation draw.

This also adds a new tool called "vipin", which is meant to parse
the telemetry strings and other information to produce a csv (or
hdf) of all worker log performance telemetry and log some basic
stats about the psimulate run.

Tested on a full SR Cervical Cancer run v6.0, and produced expected
output.