MIC-2298: Add no_cleanup option to avoid auto delete on failure
This adds a --no-cleanup option that prevents psimulate from
automatically removing the results directory. The option is aimed
at developers who may have to debug simulation failures run through
psimulate. The option is hidden (not in help output) because its
not aimed at general use.

Tested with a cancer site simulation with added bogus function
calls. Logs were preserved (or not) as expected.