MIC-2290: Add robust scenario handling, per-worker perf logs
This change adds more robust handling of scenario columns,
allowing for grid mapped scenarios to be captured in the
output. It also switches from a shared worker performance
log to using per-worker performance logs with names of the
form "perf.job_id.task_id.log".

Tested with runs of lung cancer and lsff.